added prettyPrint for arrays
added prettyPrint tests
