+++ src/test/java/org/json/simple/Test.java
@@ -374,48 +374,4 @@ public class Test extends TestCase {
 		System.out.println(jsonString);
 		assertEquals("[{\"k11\":\"v11\",\"k12\":\"v12\",\"k13\":\"v13\",\"k14\":{\"k31\":\"v3\",\"k32\":123.45,\"k33\":false,\"k34\":null,\"k35\":[\"vvv\",\"1.23456789123456789\",true,null]}},{\"k21\":\"v21\",\"k22\":\"v22\",\"k23\":\"v23\"}]", jsonString);
 	}
-
-	public void testObjectPrettyPrint() throws Exception {
-
-		JSONObject json3 = new JSONObject();
-		json3.put("ssubkey1", "ssubvalue1");
-		json3.put("ssubkey2", "ssubvalue2");
-
-		JSONObject json2 = new JSONObject();
-		json2.put("subkey1", "subvalue1");
-		json2.put("subkey2", json3);
-
-		JSONObject json1 = new JSONObject();
-
-		json1.put("key1", "value1");
-		json1.put("key2", json2);
-		json1.put("key3", json3);
-
-		System.out.println(json1.toPrettyJSONString(2));
-	}
-
-	public void testArrayPrettyPrint() throws Exception {
-
-		JSONObject json3 = new JSONObject();
-		json3.put("ssubkey1", "ssubvalue1");
-		json3.put("ssubkey2", "ssubvalue2");
-		JSONObject json2 = new JSONObject();
-		json2.put("subkey1", "subvalue1");
-		json2.put("subkey2", json3);
-		JSONObject json1 = new JSONObject();
-		json1.put("key1", "value1");
-		json1.put("key2", json2);
-		json1.put("key3", json3);
-
-		JSONArray array = new JSONArray();
-		array.add(json1);
-		array.add(json1);
-		array.add(json1);
-
-		JSONObject test = new JSONObject();
-		test.put("array", array);
-
-		System.out.println(test.toPrettyJSONString(2));
-//		System.out.println(array.toPrettyJSONString(2));
-	}
 }
